Refactor #238 use a new webpack solution based on the NODE_ENV variable
