+++ package.json
@@ -4,9 +4,8 @@
   "description": "A social network based on Symfony2 and ReactJS",
   "main": "src/Frontend/App.js",
   "scripts": {
-    "test": "mocha src/Frontend/test --recursive --ui bdd --compilers js:babel-core/register --reporter spec --require babel-polyfill --require src/Frontend/test/mocha_start.js",
-    "frontend-build": "webpack --config webpack-prod.config.js",
-    "dev-frontend-build": "webpack",
+    "test": "mocha src/Frontend/test --recursive --ui bdd --compilers js:babel-core/register --reporter dot --require babel-polyfill --require src/Frontend/test/mocha_start.js",
+    "frontend": "webpack",
     "watch": "webpack --watch",
     "lint": "eslint ./src/Frontend && lessc src/Frontend/styles/custom.less --lint"
   },
@@ -20,9 +19,9 @@
     "WebPack"
   ],
   "dependencies": {
-    "axios": "^0.8.1",
-    "babel-core": "^6",
-    "babel-loader": "^6",
+    "axios": "^0.11",
+    "babel-core": "^6.7",
+    "babel-loader": "^6.2",
     "babel-plugin-globals": "^1.0.1",
     "babel-plugin-transform-react-jsx": "^6.2.0",
     "babel-polyfill": "^6.3.14",
@@ -31,31 +30,31 @@
     "babel-preset-stage-0": "^6.1.18",
     "bootstrap": "^3.3.5",
     "cookies-js": "^1.2.2",
-    "counterpart": "^0.16.10",
+    "counterpart": "^0.17",
     "cross-spawn": "^2.0.0",
-    "css-loader": "~0.14",
-    "extract-text-webpack-plugin": "~0.8",
-    "fbjs": "^0.4",
+    "css-loader": "^0.23",
+    "extract-text-webpack-plugin": "^1.0",
+    "fbjs": "^0.8",
     "file-loader": "^0.8.4",
     "flux": "^2.1.1",
-    "history": "^1.17.0",
+    "history": "^2.1.0",
     "invariant": "^2.2.0",
-    "less": "~2.5",
+    "less": "^2.6",
     "less-loader": "^2.2",
     "mkdirp": "~0.5",
     "npmconf": "~2.1.2",
     "object.assign": "^4.0.3",
-    "perfect-scrollbar": "0.6.5-1",
-    "react": "~0.14",
+    "perfect-scrollbar": "^0.6.11",
+    "react": "^0.14",
     "react-bootstrap": "^0.28",
-    "react-dom": "~0.14",
+    "react-dom": "^0.14",
     "react-interpolate-component": "^0.8",
-    "react-router": "~1.0",
+    "react-router": "^2.0",
     "react-spinkit": "^1.1.6",
     "react-translate-component": "^0.10",
     "rewire": "^2.5.0",
-    "style-loader": "~0.8",
-    "uglify-js": "~2.4",
+    "style-loader": "^0.13",
+    "uglify-js": "^2.6",
     "url-loader": "^0.5.6",
     "webpack": "^1.12.6"
   },
@@ -67,10 +66,10 @@
   "devDependencies": {
     "babel-polyfill": "^6.2.0",
     "chai": "^3.4.1",
-    "jsdom": "^2.0.0",
+    "jsdom": "^8.4.0",
     "mocha": "^2.3.4",
     "sinon": "^1.17.2",
-    "request": "~2.65",
+    "request": "~2.72",
     "xhr2": "^0.1.3"
   }
 }
