+++ vagrant/hieradata/common.yaml
@@ -9,8 +9,10 @@ sententiaregum::installs::packages:
   - libsqlite3-dev
   - curl
 
-# ssh entry point
+# SSH config
 sententiaregum::ssh::entry_point: /var/www/sententiaregum
+sententiaregum::ssh::env:
+  NODE_ENV: development
 
 # composer config
 composer::command_name: composer
@@ -18,7 +20,7 @@ composer::target_dir: '/usr/local/bin'
 
 # php extensions
 php::service: apache
-x
+
 sententiaregum::backend::php::extensions:
   gd:     {}
   cli:    {}
