+++ vagrant/manifests/site.pp
@@ -1,4 +1,4 @@
-Exec { path => '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin' }
+Exec { path => '/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/local/node/node-default/bin' }
 Class['::apt::update'] -> Package <| |>
 
 hiera_include('classes')
