+++ vagrant/puppet/library/sententiaregum/templates/env_exporter.sh.erb
@@ -1,0 +1,6 @@
+# DON'T UPDATE THIS
+# ALL CHANGES WILL BE UNDONE IN THE NEXT PUPPET
+# PROVISIONING
+<% @env.each do |k, v| %>
+export <%= k %>=<%= v %>
+<% end %>
