+++ vagrant/puppet/library/sententiaregum/templates/vhost.erb
@@ -1,3 +1,7 @@
+# DON'T UPDATE THIS
+# ALL CHANGES WILL BE UNDONE IN THE NEXT PUPPET
+# PROVISIONING
+
 # Custom vhost for puppet
 <VirtualHost *:80>
     ServerName <%= @server_name_value %>
