--- webpack-prod.config.js
@@ -1,11 +1,0 @@
-/*
- * This file is part of the Sententiaregum project.
- *
- * (c) Maximilian Bosch <maximilian.bosch.27@gmail.com>
- * (c) Ben Bieler <benjaminbieler2014@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-module.exports = require('./webpack.factory')('prod');
